Border Control
The task is to control the border of a citizens entering a city. First we input the citizens which can be either humans or robots, In case a human enters, we have his Name, Age, and ID. If there is a robot entering, we take its Model and ID. In the end we receive one line containing the last digits of a fake IDs which we have to track and print on the console.